Eliminated compilation warning.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@19605 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/bugplug/bugplug.cpp
@@ -2316,7 +2316,7 @@ void csBugPlug::Dump (int indent, iMeshWrapper* mesh)
       	indent, "", sn ? sn : "?");
     }
   }
-  for (size_t i=0; i<mesh->GetChildren ()->GetCount (); ++i)
+  for (int i=0; i<mesh->GetChildren ()->GetCount (); ++i)
   {
     Dump (indent+4, mesh->GetChildren ()->Get (i));
   }
